Export symbols more reliable with enabled LTO
Inline virtual methods defined in headers may be removed during LTO
when they appear unused in the whole program analysis.
This affects even classes with an attribute like LIBAUDCORE_PUBLIC
because it happens before the visibility/export decisions are applied.

Moving the definitions to the respective .cc files ensures the vtable
and symbols are properly emitted and retained.

Closes: #1683